1/17/24: first pass at refactor
I kept the original logic but commented it out, and did a first pass at
a refactor. It's still not very pretty, but at least it looks better
than the original wicked conditional tree. I think reading the slightly
refactored code is a lot easier on the eyes, and gives you a clearer
idea of what happens to the item.