1/17/24: add new objects
I created 4 new classes to house the logic for each type of item (other
than Conjured, which I haven't implemented yet). This way, the 'update_quality'
method is just instantiating a certain class and relegating the logic in that
class, vs. being concerned with all the implementation in a giant case statement.
This is still not the ideal solution, but I think it's a step in the right
direction. What nudged me in this direction was reading Jim Weirich's first solution
pass at this kata. I realized what was tripping me up was instantiating a new
object with its own properties. However, for now it's ok to create a new instance
but update the initial item's properties instead. I'm guessing this is still not
ideal, but at least we are in the process of creating small objects.