1/17/24: rename class to make it more generic
If we want to add future items that have the same condition as the
Sulfuras item, i.e., the quality and sell_in remain the same, then
it'd make sense to have a more generic class that could apply to
those future items as well. Hence, I changed the name of the Sulfuras
class to Unchanged.